[OSPK8-322] Introduce ipset again to also support service VIPs
This re-introduce the ipset concept, which gives the flexibility
to request separate IPs for additional services, which is required
for OSP17 (OVNDBs, Redis). With the IPset also used for the
other resources (osclient, vmset, bmset) all of them use the
same mechanism and workflow to create hostnames and assign IPs.

With this change also alligns the scale up/down steps/procedure
from vmset and bmset, as they were not alligned.