[OSPK8-430] Allow adding additional disks to a vm role
Adds AdditionalDisks to a vmset CR which allows to specify
additional OpenStackVMSetAdditionalDisks with the parameters:
- Name
- DiskSize
- StorageClass
- StorageAccessMode
- StorageVolumeMode
- DedicatedIOThread

Also adds DedicatedIOThread for the root disk and IOThreadsPolicy and
BlockMultiQueue to the root vmset spec:

* DedicatedIOThread - Disks with dedicatedIOThread set to true will be
allocated an exclusive thread. This is generally useful if a specific
Disk is expected to have heavy I/O traffic, e.g. a database spindle.

* If IOThreadsPolicy is default, use of IOThreads will be disabled. However,
if any disk requests a dedicated IOThread, ioThreadsPolicy will be enabled
and default to shared.
When ioThreadsPolicy is set to auto IOThreads will also be "isolated" from
the vCPUs and placed on the same physical CPU as the QEMU emulator thread.
An ioThreadsPolicy of shared indicates that KubeVirt should use one thread
that will be shared by all disk devices.

* Block Multi-Queue is a framework for the Linux block layer that maps Device
I/O queries to multiple queues.
This splits I/O processing up across multiple threads, and therefor multiple
CPUs. libvirt recommends that the number of queues used should match the
number of CPUs allocated for optimal performance.